Implement filechooser
This is a request for comment, rather than a potential commit.

1. All folder/file dialogs are via gtk_file_chooser(). This provides a
consistent user interface, and the dialog itself has advantages.

2. The dialog is called from a single location, and the interface is via
a simple structure, except for the With Rename action

3. Previews of folders/some images/pdf/icc/some text files.

4. A history combo is included where sensible.

Regression:

The copy/move dialog had a "dot" as a shortcut to the current folder.
The new dialog also has a shortcut to the current folder, but its name
is the final folder name in the path - there is no way to change it to
be a "dot". It is therefore not easy to see which the current folder
shortcut is.